refactor: remove member hooks
