+++ src/services/auth/plugins/magicLink/magicLink.test.ts
@@ -6,13 +6,7 @@ import { v4 } from 'uuid';
 
 import { FastifyInstance } from 'fastify';
 
-import {
-  HttpMethod,
-  MAX_USERNAME_LENGTH,
-  MemberFactory,
-  RecaptchaAction,
-  RecaptchaActionType,
-} from '@graasp/sdk';
+import { HttpMethod, MemberFactory, RecaptchaAction, RecaptchaActionType } from '@graasp/sdk';
 import { FAILURE_MESSAGES } from '@graasp/translations';
 
 import build, { clearDatabase } from '../../../../../test/app';
@@ -22,7 +16,7 @@ import { AppDataSource } from '../../../../plugins/datasource';
 import { MailerService } from '../../../../plugins/mailer/service';
 import { AUTH_CLIENT_HOST, JWT_SECRET } from '../../../../utils/config';
 import { Member } from '../../../member/entities/member';
-import { expectMember, saveMember } from '../../../member/test/fixtures/members';
+import { saveMember } from '../../../member/test/fixtures/members';
 import { MOCK_CAPTCHA } from '../captcha/test/utils';
 
 jest.mock('node-fetch');
@@ -50,183 +44,6 @@ describe('Auth routes tests', () => {
     jest.clearAllMocks();
     await clearDatabase(app.db);
     app.close();
-  });
-
-  describe('POST /register', () => {
-    beforeEach(() => {
-      // mock captcha validation
-      mockCaptchaValidation(RecaptchaAction.SignUp);
-    });
-
-    it('Sign Up successfully', async () => {
-      const email = 'someemail@email.com';
-      const name = 'anna';
-      const mockSendEmail = jest.spyOn(mailerService, 'sendRaw');
-      const response = await app.inject({
-        method: HttpMethod.Post,
-        url: '/register',
-        payload: { email, name, captcha: MOCK_CAPTCHA },
-      });
-      const m = await memberRawRepository.findOneBy({ email, name });
-
-      expectMember(m, { name, email });
-      expect(m?.lastAuthenticatedAt).toBeNull();
-      expect(m?.isValidated).toBeFalsy();
-
-      // ensure that the user agreements are set for new registration
-      expect(m?.userAgreementsDate).toBeDefined();
-      expect(m?.userAgreementsDate).toBeInstanceOf(Date);
-
-      expect(response.statusCode).toEqual(StatusCodes.NO_CONTENT);
-      expect(mockSendEmail).toHaveBeenCalledTimes(1);
-      expect(mockSendEmail.mock.calls[0][1]).toBe(email);
-    });
-
-    it('Sign Up successfully with given lang', async () => {
-      const email = 'someemail@email.com';
-      const name = 'anna';
-      const lang = 'fr';
-
-      const mockSendEmail = jest.spyOn(mailerService, 'sendRaw');
-      const response = await app.inject({
-        method: HttpMethod.Post,
-        url: `/register?lang=${lang}`,
-        payload: { email, name, captcha: MOCK_CAPTCHA },
-      });
-
-      expect(mockSendEmail).toHaveBeenCalledTimes(1);
-      expect(mockSendEmail.mock.calls[0][1]).toBe(email);
-      const m = await memberRawRepository.findOneBy({ email, name });
-      expectMember(m, { name, email, extra: { lang } });
-      expect(m?.lastAuthenticatedAt).toBeNull();
-      expect(m?.isValidated).toBeFalsy();
-
-      // ensure that the user agreements are set for new registration
-      expect(m?.userAgreementsDate).toBeDefined();
-      expect(m?.userAgreementsDate).toBeInstanceOf(Date);
-      expect(response.statusCode).toEqual(StatusCodes.NO_CONTENT);
-    });
-
-    it('Rejects if username is too long', async () => {
-      const email = 'some@email.com';
-      const name = Array(MAX_USERNAME_LENGTH + 1).fill(() => 'a');
-
-      const response = await app.inject({
-        method: HttpMethod.Post,
-        url: `/register`,
-        payload: { email, name, captcha: MOCK_CAPTCHA },
-      });
-
-      const m = await memberRawRepository.findOneBy({ email });
-      expect(m).toBeFalsy();
-      expect(response.statusCode).toEqual(StatusCodes.BAD_REQUEST);
-    });
-
-    it('Save actions is disabled when explicitly asked', async () => {
-      const email = 'someemail@email.com';
-      const name = 'anna';
-      const enableSaveActions = false;
-
-      const mockSendEmail = jest.spyOn(mailerService, 'sendRaw');
-      const response = await app.inject({
-        method: HttpMethod.Post,
-        url: `/register`,
-        payload: { email, name, captcha: MOCK_CAPTCHA, enableSaveActions },
-      });
-
-      expect(mockSendEmail).toHaveBeenCalledTimes(1);
-      expect(mockSendEmail.mock.calls[0][1]).toBe(email);
-      const m = await memberRawRepository.findOneBy({ email, name });
-      expectMember(m, { name, email });
-      expect(m?.enableSaveActions).toBe(enableSaveActions);
-      // ensure that the user agreements are set for new registration
-      expect(m?.userAgreementsDate).toBeDefined();
-      expect(m?.userAgreementsDate).toBeInstanceOf(Date);
-      expect(m?.lastAuthenticatedAt).toBeNull();
-      expect(m?.isValidated).toBeFalsy();
-
-      expect(response.statusCode).toEqual(StatusCodes.NO_CONTENT);
-    });
-
-    it('Enable save actions when explicitly asked', async () => {
-      const email = 'someemail@email.com';
-      const name = 'anna';
-      const enableSaveActions = true;
-
-      const mockSendEmail = jest.spyOn(mailerService, 'sendRaw');
-      const response = await app.inject({
-        method: HttpMethod.Post,
-        url: `/register`,
-        payload: { email, name, enableSaveActions, captcha: MOCK_CAPTCHA },
-      });
-
-      expect(mockSendEmail).toHaveBeenCalledTimes(1);
-      expect(mockSendEmail.mock.calls[0][1]).toBe(email);
-      const m = await memberRawRepository.findOneBy({ email, name });
-      expectMember(m, { name, email });
-      expect(m?.enableSaveActions).toBe(enableSaveActions);
-      // ensure that the user agreements are set for new registration
-      expect(m?.userAgreementsDate).toBeDefined();
-      expect(m?.userAgreementsDate).toBeInstanceOf(Date);
-      expect(m?.lastAuthenticatedAt).toBeNull();
-      expect(m?.isValidated).toBeFalsy();
-
-      expect(response.statusCode).toEqual(StatusCodes.NO_CONTENT);
-    });
-
-    it('Sign Up fallback to login for already register member', async () => {
-      // register already existing member
-      const member = await saveMember(MemberFactory({ isValidated: false }));
-      const mockSendEmail = jest.spyOn(mailerService, 'sendRaw');
-
-      const response = await app.inject({
-        method: HttpMethod.Post,
-        url: '/register',
-        payload: { ...member, captcha: MOCK_CAPTCHA },
-      });
-
-      expect(mockSendEmail).toHaveBeenCalledTimes(1);
-      expect(mockSendEmail.mock.calls[0][1]).toBe(member.email);
-
-      const members = await memberRawRepository.findBy({ email: member.email });
-      expect(members).toHaveLength(1);
-      expectMember(member, members[0]);
-      expect(members[0]?.lastAuthenticatedAt).toBeNull();
-      expect(members[0]?.isValidated).toBeFalsy();
-      expect(response.statusCode).toEqual(StatusCodes.NO_CONTENT);
-    });
-
-    it('Bad request for invalid email', async () => {
-      const email = 'wrongemail';
-      const name = 'anna';
-      const response = await app.inject({
-        method: HttpMethod.Post,
-        url: '/register',
-        payload: { email, name, captcha: MOCK_CAPTCHA },
-      });
-
-      const members = await memberRawRepository.findBy({ email });
-      expect(members).toHaveLength(0);
-
-      expect(response.statusMessage).toEqual(ReasonPhrases.BAD_REQUEST);
-      expect(response.statusCode).toEqual(StatusCodes.BAD_REQUEST);
-    });
-
-    it('Bad request if the username contains special characters', async () => {
-      const email = faker.internet.email().toLowerCase();
-      const name = '<div>%"^';
-      const response = await app.inject({
-        method: HttpMethod.Post,
-        url: '/register',
-        payload: { email, name, captcha: MOCK_CAPTCHA },
-      });
-
-      const members = await memberRawRepository.findBy({ email });
-      expect(members).toHaveLength(0);
-
-      expect(response.statusMessage).toEqual(ReasonPhrases.BAD_REQUEST);
-      expect(response.statusCode).toEqual(StatusCodes.BAD_REQUEST);
-    });
   });
 
   describe('POST /login', () => {
